requirements.txt
annotated-types==0.7.0
packaging==26.1
pyaes==1.6.1
pyasn1==0.6.3
pydantic==2.13.2
pydantic_core==2.46.2
python-dotenv==1.2.2
rsa==4.9.1
setuptools==82.0.1
Telethon==1.43.0
typing-inspection==0.4.2
typing_extensions==4.15.0
wheel==0.46.3
